pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "googlesearch-python>=1.3.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
